+++ Source/Utility/MythForest/Component/Camera/CameraComponent.cpp
@@ -188,8 +188,8 @@ void CameraComponent::Instancing(Engine& engine, TaskData& taskData) {
 					assert(group.drawCallDescription.bufferResources[output.slot].buffer == nullptr);
 					IRender::Resource::DrawCallDescription::BufferRange& bufferRange = group.drawCallDescription.bufferResources[output.slot];
 					bufferRange.buffer = policyData.instanceBuffer;
-					bufferRange.offset = policyData.instanceData.GetSize(); // TODO: alignment
-					bufferRange.component = data.GetSize() / (group.instanceCount * sizeof(float));
+					bufferRange.offset = safe_cast<uint32_t>(policyData.instanceData.GetSize()); // TODO: alignment
+					bufferRange.component = safe_cast<uint32_t>(data.GetSize()) / (group.instanceCount * sizeof(float));
 					policyData.instanceData.Append(data);
 				}
 			}
